refactor copy for boom to zq
