+++ src/css/_curl-modal.scss
@@ -11,6 +11,6 @@
   }
 }
 
-.boom-get-modal {
+.zq-get-modal {
   min-height: 200px;
 }
